Refactor rendering system WIP: clean up code

+++ InnocenceEngine 1.0/source/code/game/InnocenceGarden.h
@@ -51,10 +51,10 @@ private:
 
 	std::vector<LightComponent> pointLightComponents;
 
-	double temp = 0.0f;
+	float temp = 0.0f;
 
 	void initSpheres();
 	void initLights();
-	void updateLights(double seed);
+	void updateLights(float seed);
 };
